requirements.txt
discord.py==2.3.2
python-dotenv==1.0.0
Flask==3.0.0
requests-oauthlib==1.3.1
requests==2.31.0
